chore: refactor overview out of inReviewOverview and remove the custom component
