+++ libs/application/templates/accident-notification/src/fields/index.ts
@@ -1,5 +1,4 @@
 export { DateOfAccident } from './DateOfAccident'
-export { FormOverview } from './FormOverview'
 export { HiddenInformation } from './HiddenInformation'
 export { ApplicationStatus } from './ApplicationStatus'
 export { FormOverviewInReview } from './FormOverviewInReview'
