+++ libs/application/templates/accident-notification/src/forms/InReviewForm/inReviewOverviewSection.ts
@@ -1,20 +1,32 @@
-import { buildCustomField, buildSection } from '@island.is/application/core'
+import {
+  buildCustomField,
+  buildMultiField,
+  buildSection,
+} from '@island.is/application/core'
 import { overview } from '../../lib/messages'
+import { overviewFields } from '../../utils/overviewFields'
 
 export const inReviewOverviewSection = (isAssignee?: boolean) =>
   buildSection({
     id: 'InReviewOverviewSection',
     title: overview.general.sectionTitle,
     children: [
-      buildCustomField(
-        {
-          id: 'inReviewOverviewScreen',
-          title: '',
-          component: 'FormOverviewInReview',
-        },
-        {
-          isAssignee,
-        },
-      ),
+      buildMultiField({
+        id: 'inReviewOverviewScreen',
+        title: 'Yfirlit',
+        children: [
+          ...overviewFields(isAssignee),
+          buildCustomField(
+            {
+              id: 'inReviewOverviewScreen',
+              title: '',
+              component: 'FormOverviewInReview',
+            },
+            {
+              isAssignee,
+            },
+          ),
+        ],
+      }),
     ],
   })
