+++ libs/application/templates/accident-notification/src/lib/messages/overview.ts
@@ -66,9 +66,9 @@ export const overview = {
       description: 'Description of alert message that appears on overview page',
     },
     descriptionWithFiles: {
-      id: 'an.application:overview.alertMessage.descriptionWithFiles',
+      id: 'an.application:overview.alertMessage.descriptionWithFiles#markdown',
       defaultMessage:
-        'Áður en að Sjúkratryggingar Íslands fara yfir umsókn og taka afstöðu til bótaskyldu þurfa að minnsta kosti eftirtalin fylgiskjöl að berast: **Áverkavottorð**, **Umboð** og **Lögregluskýrsla**',
+        'Áður en að Sjúkratryggingar Íslands fara yfir umsókn og taka afstöðu til bótaskyldu þurfa að minnsta kosti eftirtalin fylgiskjöl að berast: **{missingFiles}**',
       description: 'Description of alert message that appears on overview page',
     },
   }),
